backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
httpx>=0.27
pyyaml>=6.0
python-multipart>=0.0.9
trafilatura>=1.12
beautifulsoup4>=4.12
pymupdf>=1.24
numpy>=1.26
pydantic>=2.7
lxml>=5.0
feedparser>=6.0
gitpython>=3.1
Pillow>=10.0
pytesseract>=0.3
flashrank>=0.2
defusedxml>=0.7
pytest>=8.0
pytest-asyncio>=0.23
